Refactor Cargo.toml and add dev-dependencies
