+++ Cargo.toml
@@ -13,6 +13,9 @@ serde_json = "1.0"
 [build-dependencies]
 slint-build = "1.7.2"
 
+[dev-dependencies]
+i-slint-backend-testing = "1.8.0"
+
 [package.metadata.bundle]
 name = "My Finance"
 identifier = "io.github.Qubi0-0.My-Finance"
